requirements.txt
jupyterlab
openai
python-dotenv
